Supreme Secondary Learner Government of Sagkahan National High School
This file contains the official seal of the SSLG-SNHS. This file is under license with the Creative Commons License CC BY-NC-ND 4.0 
This can be used and redistributed. However, tampering, adaptation, or other such act that could compromise the 
Official Learner Government of SNHS' Seal and Coat of Arms are prohibited under the said license.